chore: Refactor node.rs into smaller files
